backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Database
sqlalchemy==2.0.23
databases[sqlite]==0.8.0
alembic==1.12.1
motor==3.3.2
pymongo==4.6.0
beanie==1.23.6

# File processing and document analysis
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
spacy==3.7.2
dateparser==1.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# Reporting and visualization
matplotlib==3.8.2
reportlab==4.0.7

# Utilities
aiofiles==23.2.1
jinja2==3.1.2
pydantic==2.5.1
pydantic[email]==2.5.0

# Phase 3: NLP and Validation dependencies
textdistance==4.6.2

# Async processing and background tasks
celery==5.3.4
redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
python-json-logger==2.0.7

# Environment and configuration
python-dotenv==1.0.0 